+++ packages/sdk/test/create.test.ts
@@ -1,13 +1,10 @@
 import {
-  App,
   Chains,
   ERC20CreateParams,
   ERC721CreateParams,
-  Errors,
   OpenFormatSDK,
 } from '../src';
 import { toWei } from '../src/helpers';
-import { mockLowFeeBalance } from './mocks/lowFeeBalance';
 
 describe('NFT', () => {
   it('creates an NFT contract and returns an instance of it', async () => {
@@ -66,7 +63,7 @@ describe('NFT', () => {
 
     const sdk = new OpenFormatSDK({
       network: Chains.foundry,
-      starId: global.star,
+      appId: global.app,
     });
 
     async function create() {
@@ -74,27 +71,6 @@ describe('NFT', () => {
     }
 
     await expect(create).rejects.toThrow('Signer undefined');
-  });
-
-  it('should fail to create an NFT when the account has insufficient funds for transaction fees and throw a low balance error', async () => {
-    mockLowFeeBalance();
-
-    const AppInstance = new App(
-      global.sdk.provider,
-      global.star,
-      global.sdk.signer
-    );
-
-    const params: ERC721CreateParams = {
-      name: 'TEST',
-      symbol: 'NFT',
-      royaltyRecipient: global.walletAddress,
-      royaltyBps: 1000,
-    };
-
-    await expect(AppInstance.createNFT(params)).rejects.toThrow(
-      Errors.LowTransactionFeeBalance
-    );
   });
 });
 
@@ -155,7 +131,7 @@ describe('NFTDrop', () => {
 
     const sdk = new OpenFormatSDK({
       network: Chains.foundry,
-      starId: global.star,
+      appId: global.app,
     });
 
     async function create() {
@@ -163,27 +139,6 @@ describe('NFTDrop', () => {
     }
 
     await expect(create).rejects.toThrow('Signer undefined');
-  });
-
-  it('should fail to create an NFTDrop when the account has insufficient funds for transaction fees and throw a low balance error', async () => {
-    mockLowFeeBalance();
-
-    const AppInstance = new App(
-      global.sdk.provider,
-      global.star,
-      global.sdk.signer
-    );
-
-    const params: ERC721CreateParams = {
-      name: 'TEST',
-      symbol: 'NFT',
-      royaltyRecipient: global.walletAddress,
-      royaltyBps: 1000,
-    };
-
-    await expect(AppInstance.createNFTDrop(params)).rejects.toThrow(
-      Errors.LowTransactionFeeBalance
-    );
   });
 });
 
@@ -231,7 +186,7 @@ describe('Token', () => {
 
     const sdk = new OpenFormatSDK({
       network: Chains.foundry,
-      starId: global.star,
+      appId: global.app,
     });
 
     async function create() {
@@ -239,25 +194,5 @@ describe('Token', () => {
     }
 
     await expect(create).rejects.toThrow('Signer undefined');
-  });
-
-  it('should fail to create a Token when the account has insufficient funds for transaction fees and throw a low balance error', async () => {
-    mockLowFeeBalance();
-
-    const AppInstance = new App(
-      global.sdk.provider,
-      global.star,
-      global.sdk.signer
-    );
-
-    const params: ERC20CreateParams = {
-      name: 'TEST',
-      symbol: 'NFT',
-      supply: 1000,
-    };
-
-    await expect(AppInstance.createToken(params)).rejects.toThrow(
-      Errors.LowTransactionFeeBalance
-    );
   });
 });
