+++ packages/sdk/test/setup.ts
@@ -4,7 +4,7 @@ require('ts-node').register({
 
 import { faker } from '@faker-js/faker';
 import { ContractFactory, ethers } from 'ethers';
-import { Chains, OpenFormatSDK, toWei, txFeeContractAddress } from '../src';
+import { Chains, OpenFormatSDK, toWei } from '../src';
 import { ERC20Base__factory } from '../src/contract-types';
 import { WALLETS } from './utilities';
 import ERC20Base from './utils/ERC20Base.json';
@@ -16,7 +16,7 @@ export const PRIVATE_KEY =
 module.exports = async () => {
   const sdk = new OpenFormatSDK({
     network: Chains.foundry,
-    starId: '',
+    appId: '',
     signer: PRIVATE_KEY,
   });
 
@@ -31,43 +31,23 @@ module.exports = async () => {
   const OPEN = await ERC20BaseFactory.deploy();
   await OPEN.deployed();
 
-  txFeeContractAddress[31337] = { address: OPEN.address };
-
   if (sdk.signer) {
     global.walletAddress = await sdk.signer?.getAddress();
   }
 
-  // Create constellation
-  const constellationTx = await sdk.factory.createConstellation({
-    name: faker.internet.domainName(),
-    symbol: faker.hacker.abbreviation(),
-    decimals: 18,
-    supply: toWei('100'),
-  });
-
-  global.constellation = constellationTx.constellationAddress;
-
-  // Create star
-  const starTx = await sdk.factory.createStar({
+  // Create app
+  const app = await sdk.factory.createApp({
     name: faker.internet.domainWord(),
-    constellation: global.constellation,
     owner: global.walletAddress,
   });
 
-  global.star = starTx.starAddress;
+  global.app = app.id;
 
   global.sdk = new OpenFormatSDK({
     network: Chains.foundry,
-    starId: global.star,
+    appId: global.app,
     signer: PRIVATE_KEY,
   });
-
-  const constellationInstance = ERC20Base__factory.connect(
-    constellationTx.constellationAddress,
-    global.sdk.signer
-  );
-
-  await constellationInstance.approve(global.star, ethers.constants.MaxUint256);
 
   // Create NFT
   const NFT = await global.sdk.App.createNFT({
@@ -92,10 +72,24 @@ module.exports = async () => {
     supply: toWei('100'),
   });
 
+  const RewardToken = await global.sdk.App.createToken({
+    name: 'REWARDTOKEN',
+    symbol: 'REWARDTOKEN',
+    supply: toWei('100'),
+  });
+
+  const rewardTokenInstance = ERC20Base__factory.connect(
+    RewardToken.address(),
+    global.sdk.signer
+  );
+
+  await rewardTokenInstance.approve(global.app, ethers.constants.MaxUint256);
+
   global.NFT = NFT;
   global.NFTDrop = NFTDrop;
   global.Token = Token;
   global.OPEN = OPEN;
+  global.RewardToken = RewardToken;
 
   console.log('Setup complete 🚀');
 };
