+++ packages/sdk/test/subgraph.test.ts
@@ -5,7 +5,7 @@ describe('subgraph', () => {
   it('should return subgraph endpoint for Polygon Mumbai', async () => {
     const sdk = new OpenFormatSDK({
       network: Chains.polygonMumbai,
-      starId: global.star,
+      appId: global.app,
     });
 
     const endpoint = await sdk.getSubgraphEndpoint();
@@ -16,7 +16,7 @@ describe('subgraph', () => {
   it('should return subgraph endpoint for Polygon Mainnet', async () => {
     const sdk = new OpenFormatSDK({
       network: Chains.polygon,
-      starId: global.star,
+      appId: global.app,
     });
 
     const endpoint = await sdk.getSubgraphEndpoint();
@@ -26,7 +26,7 @@ describe('subgraph', () => {
   it('should return subgraph endpoint for Aurora', async () => {
     const sdk = new OpenFormatSDK({
       network: Chains.aurora,
-      starId: global.star,
+      appId: global.app,
     });
 
     const endpoint = await sdk.getSubgraphEndpoint();
@@ -37,7 +37,7 @@ describe('subgraph', () => {
   it('should return subgraph endpoint for Aurora testnet', async () => {
     const sdk = new OpenFormatSDK({
       network: Chains.auroraTestnet,
-      starId: global.star,
+      appId: global.app,
     });
 
     const endpoint = await sdk.getSubgraphEndpoint();
@@ -47,7 +47,7 @@ describe('subgraph', () => {
   it('should return subgraph endpoint for Foundry(localhost)', async () => {
     const sdk = new OpenFormatSDK({
       network: Chains.foundry,
-      starId: global.star,
+      appId: global.app,
     });
 
     const endpoint = await sdk.getSubgraphEndpoint();
@@ -58,12 +58,12 @@ describe('subgraph', () => {
   it('allows you to perform a raw request', async () => {
     const sdk = new OpenFormatSDK({
       network: Chains.polygonMumbai,
-      starId: global.star,
+      appId: global.app,
     });
 
     const query = gql`
       {
-        constellations {
+        users {
           id
         }
       }
@@ -71,6 +71,6 @@ describe('subgraph', () => {
 
     const result = await sdk.subgraph.rawRequest(query);
 
-    expect(typeof result.constellations[0].id).toBe('string');
+    expect(typeof result.users[0].id).toBe('string');
   });
 });
